requirements.txt
pytz>=2020.4
aiohttp>=3.8.1,<4
requests>=2.28.1,<3
pyjwt
python-socketio>=5.9,<6

